Add a notion of event scripts
Event scripts are configurable at any time with reference to EngineState and contain lists of commands to execute in response to certain events.
Rework a few behaviors to work as engine scripts instead of hard-coded messages
Remove NotFoundCommandVerb, it's not needed. That behavior is part of a script now
Add an Echo verb so that we can tell the console to output values