Start working on renames
s/ICommandVerb/ICommandHandler/
s/ICommandVerbSource/ICommandHandlerSource/
a verb is just the string name. A command is a verb with arguments. A handler is what is invoked to handle the command
This is despite the fact that the handlers are implementations of the Command Pattern. We need the rename for clarity between separate concepts in our library