Rewrite argument handling
Differentiate between the raw parsed argument IParsedArgument and the accessible IArgument
Ambiguous cases in argument grammars can be resolved downstream by seeing user intent. When an arg is accessed, we know unambiguously what it was supposed to be.
In these case we put in a combined IParsedArgument, and when it's accessed we can replace it by individual arguments
Several TODO notes, especially for the performance of multiple list traversals (which might be negligible, we're probably dealing with small lists here)