EngineSettings object and configurable loop limit
Institute a mechanism to count the number of commands executed without user input and stop execution if the number goes above a configurable limit.
Default value is 20, but settable in EngineBuilder
Add a new MaximumHeadlessCommands script which is executed to alert the user about the loop and exit the engine
In interactive mode the user is prompted to continue ('y') or break ('n') with a break causing the current queue to clear.
+integration tests