Store the current Command in the EngineState temporarily
We only need it to be there for a few ticks while we invoke the IHandlerSource to get the handler, so the DI container can resolve the handler with the command/args. After resolution, the contents of this property does not matter.
This change allows us to remove all resolve-time temporary registrations for Lamar and StructureMap handler sources
misc cleanups