Remove the exec and run handlers
Neither of these are used internally.
run wasn't useful by itself, there are no places in the code where we need to script a sequence of commands and cannot use multiple strings for the purpose.
exec wasn't useful because it would only search the current working dir instead of a more useful place (or list of places), and it's probably better in most cases to script from the shell instead of from inside stonefruit