Start switching over to HostApplicationBuilder
Instead of having code to set up all sorts of different DI containers, we're going to piggy-back on the .NET Generic Application Host (IHostApplicationBuilder) and simply take a constructed IServiceProvider as set up by that system.
Remove Scrutor dependency. It is not reliable and I have had several problems with it in downstream usages
Several TODO notes, including needing to re-implement scanning and resolution of handler types