Fix ServiceProviderHandlerSource to not use IServiceCollection
We don't know the lifetime of IServiceCollection and it is absolutely NOT established idiom to keep it around just in case. A memory-conscious programmer could easily clear and/or dispose it before we ever had the chance to inspect it's contents.
Instead register a RegisteredHandler object with the handler's Type. Then we can inject the list of those and do all our reflection that way
Simplifies a lot of code and removes a huge source of possible bugs