Add a HandlerContext to hold information needed to find and execute a context
CommandDispatcher used to be used as a mule to expose a bunch of values, since it had them injected already. Now HandlerContext is the short-lived mule for those values and has no other state/functionality
IHandlers and IHandlerSource now take HandlerContext in their Get() methods
IHandlerMethodInvoker now takes the HandlerContext instead of args and dispatcher
Added two new IHandlerBase variants: IHandlerWithContext and IHandlerAsyncWithContext. These allow the CommandDispatcher to inject the HandlerContext directly
HandlerContext IS NOT injected into the DI. We could try to do this in the future (probably replacing EngineState.CurrentArguments), but right now we don't have a use-case
Several other cleanups and improvements throughout