Value-type parsing during dynamic method invoke
If we have a reference type in the IHandlerMethodInvoker, we can see if there is a registered IValueTypeMapper<T> for that type. Then we can pass our IValuedArgument to the mapper and get back the mapped value instance
Also rename IParserSetup and friends to ArgumentSetup, since we're talking about all sorts of argument parsing/mappingThis only works for public instance methods. The feature currently is not exposed anywhere else.