IServiceCollection.AddHandlerArgumentType<>() method
This method allows us to register an object as an arguments type. Then we can inject pre-mapped argument values anywhere that we can inject other dependencies and not have to do our own argument mapping/decoding
Also respect the [Required] annotation while mapping. If a required property hasn't been mapped, an error is thrown.
There are a few other annotations from System.ComponentModel.DataAnnotations that I think we should support in the future